requirements.txt
aiohttp==3.7.4.post0
aioredis==1.3.1
amqp==5.0.6
asgiref==3.4.1
async-timeout==3.0.1
atlastk==0.12.10
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
autopep8==1.5.7
backcall==0.2.0
backoff==1.11.1
beautifulsoup4==4.9.3
billiard==3.6.4.0
bs4==0.0.1
CacheControl==0.12.6
celery==5.1.2
certifi==2021.5.30
cffi==1.14.6
channels==3.0.4
channels-redis==3.3.0
chardet==4.0.0
charset-normalizer==2.0.3
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
cloudinary==1.26.0
colorama==0.4.4
config==0.5.0.post0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
daphne==3.0.2
decorator==4.4.2
decouple==0.0.7
defusedxml==0.7.1
diff-match-patch==20200713
dill==0.3.4
dj-database-url==0.5.0
Django==3.2.7
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-cleanup==5.2.0
django-cors-headers==3.7.0
django-crispy-forms==1.12.0
django-ffmpeg==0.1.4
django-filter==2.4.0
django-import-export==2.5.0
django-jazzmin==2.4.8
django-jet-reboot==1.1.1
django-rest-swagger==2.2.0
django-sslify==0.2.7
django-timezone-field==4.2.1
djangorestframework==3.12.4
dnspython==2.1.0
docx2pdf==0.1.7
drf-yasg==1.20.0
et-xmlfile==1.1.0
eventlet==0.32.0
Faker==8.13.2
feedparser==6.0.8
greenlet==1.1.1
gunicorn==20.1.0
hiredis==2.0.0
html5lib==1.1
httpie==2.4.0
hyperlink==21.0.0
idna==3.2
imageio==2.9.0
imageio-ffmpeg==0.4.4
incremental==21.3.0
inflection==0.5.1
ipython==7.27.0
itypes==1.2.0
jedi==0.18.0
Jinja2==3.0.1
kombu==5.1.0
MarkupPy==1.14
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
moviepy==1.0.3
msgpack==1.0.2
multidict==5.1.0
multiprocess==0.70.12.2
mysqlclient==2.0.3
numpy==1.21.1
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.0.7
packaging==21.0
pafy==0.5.5
pandas==1.3.2
parser201==0.1.7
parso==0.8.2
pickleshare==0.7.5
Pillow==8.3.1
proglog==0.1.9
prompt-toolkit==3.0.20
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pyfy==2.0.9
Pygments==2.9.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
PySocks==1.7.1
python-crontab==2.5.1
python-dateutil==2.8.2
python-decouple==3.4
pytube==11.0.0
pytz==2021.1
pywin32==227
PyYAML==5.4.1
redis==3.5.3
requests==2.26.0
requests-toolbelt==0.9.1
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
selenium==3.141.0
service-identity==21.1.0
sgmllib3k==1.0.0
simplejson==3.17.5
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
tablib==3.0.0
text-unidecode==1.3
toml==0.10.2
tortoise==0.1.1
tqdm==4.62.0
traitlets==5.1.0
Twisted==21.7.0
twisted-iocpsupport==1.0.2
txaio==21.2.1
typing-extensions==3.10.0.0
Unipath==1.1
uritemplate==3.0.1
urllib3==1.26.6
vine==5.0.0
wadofstuff-django-serializers==1.1.0
wcwidth==0.2.5
webencodings==0.5.1
wget==3.2
whitenoise==5.3.0
xlrd==2.0.1
xlwt==1.3.0
yarl==1.6.3
youtube-dl==2021.6.6
zope.interface==5.4.0
